+++ pyproject.toml
@@ -39,7 +39,9 @@ dependencies = [
 
 [project.optional-dependencies]
 test = [
-    "pytest~=7.1"
+    "pytest~=7.1",
+    "pytest-mock~=3.10",
+    "pytest-asyncio~=0.18"
 ]
 
 lint = ["pylint~=2.14"]
